backend/requirements.txt
FlagEmbedding
torch
transformers
spicy
qdrant_client
python-dotenv
pandas
openai
instructor